Moved player list handling to ItemDropHandler for Items Rewards, and AlertPlayers. Added to Raid function as well. Fix to weird message sending.
