+++ Packages/manifest.json
@@ -3,20 +3,20 @@
     "com.unity.burst": "1.7.0-pre.1",
     "com.unity.cloud.userreporting": "0.2.4-preview",
     "com.unity.collab-proxy": "1.13.5",
-    "com.unity.ide.rider": "3.0.7",
-    "com.unity.ide.visualstudio": "2.0.12",
+    "com.unity.ide.rider": "3.0.12",
+    "com.unity.ide.visualstudio": "2.0.14",
     "com.unity.ide.vscode": "1.2.4",
     "com.unity.mathematics": "1.2.5",
-    "com.unity.memoryprofiler": "0.4.2-preview.1",
+    "com.unity.memoryprofiler": "0.4.4-preview.2",
     "com.unity.performance.profile-analyzer": "1.1.1",
     "com.unity.polybrush": "1.1.2",
     "com.unity.postprocessing": "3.1.1",
     "com.unity.render-pipelines.universal": "12.1.0",
     "com.unity.searcher": "4.9.1",
-    "com.unity.test-framework": "1.1.29",
+    "com.unity.test-framework": "1.1.30",
     "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.6.2",
-    "com.unity.toolchain.win-x86_64-linux-x86_64": "0.1.21-preview",
+    "com.unity.timeline": "1.6.4",
+    "com.unity.toolchain.win-x86_64-linux-x86_64": "1.0.0",
     "com.unity.ugui": "1.0.0",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
